@@ -64,7 +64,7 @@ testpaths = [
 [tool.ruff]
 line-length = 88 # black's default
 src = ["src", "tests"]
-target-version = "py38"
+target-version = "py313"
 
 [tool.ruff.format]
 docstring-code-format = true
