@@ -15,7 +15,7 @@ classifiers =
     Framework :: Flask
     License :: OSI Approved :: Apache Software License
     Programming Language :: Python
-    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.13
     Topic :: Internet :: WWW/HTTP :: HTTP Servers
 
 [options]
@@ -23,7 +23,7 @@ zip_safe = True
 packages = find:
 package_dir =
     =src
-python_requires = >=3.8
+python_requires = >=3.13
 include_package_data = True
 install_requires =
     flask>=2.2
